bank

+++ Account.java
@@ -2,18 +2,18 @@ import java.util.Random;
 
 public class Account {
     private int numberAccount;
-    private int accountBalance;
+    private double accountBalance;
 
     public Account(){
         this.numberAccount = (int) Math.random() * 999 + 1;
         this.accountBalance = 0;
     }
-    public Account(int accountBalance){
+    public Account(double accountBalance){
         this.accountBalance = accountBalance;
         this.numberAccount = (int) Math.random() * 999 + 1;
     }
 
-    public void setBalance(int balance){
+    public void setBalance(double balance){
         this.accountBalance = balance;
     }
     public double getBalance(){
